Add documentation of InsertFuture
