+++ docs/source/api/search.rst
@@ -112,37 +112,3 @@ APIs References
    :member-order: bysource
    :members: id, distance, score, __str__
 
-
-SearchResultFuture
-------------------
-
-Constructor
-~~~~~~~~~~~
-
-+----------------------------------------------------------------------+------------------------------------------------------------------------+
-| Constructor                                                          | Description                                                            |
-+======================================================================+========================================================================+
-| `SearchResultFuture() <#pymilvus_orm.SearchResultFuture>`_           | Search result future.                                    |
-+----------------------------------------------------------------------+------------------------------------------------------------------------+
-
-Attributes
-~~~~~~~~~~
-
-+----------------------------------------------------------------------+------------------------------------------------------------------------+
-| API                                                                  | Description                                                            |
-+======================================================================+========================================================================+
-| `result() <#pymilvus_orm.SearchResultFuture.result>`_                | Return the search result.                                              |
-+----------------------------------------------------------------------+------------------------------------------------------------------------+
-| `cancel() <#pymilvus_orm.SearchResultFuture.cancel>`_                | Cancel the search request.                                             |
-+----------------------------------------------------------------------+------------------------------------------------------------------------+
-| `done() <#pymilvus_orm.SearchResultFuture.done>`_                    | Wait for search request done.                                          |
-+----------------------------------------------------------------------+------------------------------------------------------------------------+
-
-
-APIs References
-~~~~~~~~~~~~~~~
-
-
-.. autoclass:: pymilvus_orm.SearchResultFuture
-   :member-order: bysource
-   :members: result, cancel, done
